+++ android-sdk/src/main/java/io/relayr/SocketClient.java
@@ -1,5 +1,6 @@
 package io.relayr;
 
+import io.relayr.model.Reading;
 import io.relayr.model.TransmitterDevice;
 import rx.Observable;
 
@@ -9,7 +10,7 @@ public interface SocketClient {
      * Subscribes an app to a device channel. Enables the app to receive data from the device.
      * @param device The device object to be subscribed to.
      */
-    public Observable<Object> subscribe(TransmitterDevice device);
+    public Observable<Reading> subscribe(TransmitterDevice device);
 
     /**
      * Unsubscribes an app from a device channel, stopping and cleaning up the connection.