+++ android-sdk/src/main/java/io/relayr/ble/parser/BleDataParser.java
@@ -4,8 +4,9 @@ import com.google.gson.Gson;
 
 import io.relayr.ble.BleDeviceType;
 import io.relayr.model.AccelGyroscope;
-import io.relayr.model.BleReading;
+import io.relayr.model.DeviceModel;
 import io.relayr.model.LightColorProx;
+import io.relayr.model.DataPackage;
 
 public abstract class BleDataParser {
 
@@ -29,18 +30,30 @@ public abstract class BleDataParser {
     }
 
     private static String getLIGHTSensorData(byte[] value) {
-        BleReading reading = new BleReading();
-        reading.readings.color = new LightColorProx.Color();
-        reading.readings.luminosity = (byteToUnsignedInt(value[1]) << 8) | byteToUnsignedInt(value[0]);
-        reading.readings.color.red = (byteToUnsignedInt(value[3]) << 8) | byteToUnsignedInt(value[2]);
-        reading.readings.color.green = (byteToUnsignedInt(value[5]) << 8) | byteToUnsignedInt(value[4]);
-        reading.readings.color.blue = (byteToUnsignedInt(value[7]) << 8) | byteToUnsignedInt(value[6]);
-        reading.readings.proximity = (byteToUnsignedInt(value[9]) << 8) | byteToUnsignedInt(value[8]);
-        return new Gson().toJson(reading);
+        DataPackage dataPackage = new DataPackage();
+        dataPackage.modelId = DeviceModel.LIGHT_PROX_COLOR.getId();
+        dataPackage.received = System.currentTimeMillis();
+
+        int red = (byteToUnsignedInt(value[3]) << 8) | byteToUnsignedInt(value[2]);
+        int green = (byteToUnsignedInt(value[5]) << 8) | byteToUnsignedInt(value[4]);
+        int blue = (byteToUnsignedInt(value[7]) << 8) | byteToUnsignedInt(value[6]);
+        LightColorProx.Color color = new LightColorProx.Color(red, green, blue);
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "color", "", color));
+
+        int proximity = (byteToUnsignedInt(value[9]) << 8) | byteToUnsignedInt(value[8]);
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "proximity", "", proximity));
+        
+        int luminosity = (byteToUnsignedInt(value[1]) << 8) | byteToUnsignedInt(value[0]);
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "luminosity", "", luminosity));
+        
+        return new Gson().toJson(dataPackage);
     }
 
     private static String getGYROSensorData(byte[] value) {
-        BleReading reading = new BleReading();
+        DataPackage dataPackage = new DataPackage();
+        dataPackage.modelId = DeviceModel.ACCELEROMETER_GYROSCOPE.getId();
+        dataPackage.received = System.currentTimeMillis();
+        
         int gyroscopeX = byteToUnsignedInt(value[0]) |
                 (byteToUnsignedInt(value[1]) << 8) |
                 (byteToUnsignedInt(value[2]) << 16) |
@@ -58,32 +71,43 @@ public abstract class BleDataParser {
         int accelerationY = (byteToUnsignedInt(value[15]) << 8) | byteToUnsignedInt(value[14]);
         int accelerationZ = (byteToUnsignedInt(value[17]) << 8) | byteToUnsignedInt(value[16]);
 
-        reading.readings.acceleration = new AccelGyroscope.Accelerometer();
-        reading.readings.acceleration.x = (float) accelerationX / 100.0f;
-        reading.readings.acceleration.y = (float) accelerationY / 100.0f;
-        reading.readings.acceleration.z = (float) accelerationZ / 100.0f;
+        AccelGyroscope.Acceleration acceleration = new AccelGyroscope.Acceleration();
+        acceleration.x = (float) accelerationX / 100.0f;
+        acceleration.y = (float) accelerationY / 100.0f;
+        acceleration.z = (float) accelerationZ / 100.0f;
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "acceleration", "", acceleration));
 
-        reading.readings.angularSpeed = new AccelGyroscope.Gyroscope();
-        reading.readings.angularSpeed.x = (float) gyroscopeX / 100.0f;
-        reading.readings.angularSpeed.y = (float) gyroscopeY / 100.0f;
-        reading.readings.angularSpeed.z = (float) gyroscopeZ / 100.0f;
-        return new Gson().toJson(reading);
+        AccelGyroscope.AngularSpeed angularSpeed = new AccelGyroscope.AngularSpeed();
+        angularSpeed.x = (float) gyroscopeX / 100.0f;
+        angularSpeed.y = (float) gyroscopeY / 100.0f;
+        angularSpeed.z = (float) gyroscopeZ / 100.0f;
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "angularSpeed", "", angularSpeed));
+        
+        return new Gson().toJson(dataPackage);
     }
 
     private static String getHTUSensorData(byte[] value) {
-        BleReading reading = new BleReading();
+        DataPackage dataPackage = new DataPackage();
+        dataPackage.modelId = DeviceModel.TEMPERATURE_HUMIDITY.getId();
+        dataPackage.received = System.currentTimeMillis();
+
         int temperature = (byteToUnsignedInt(value[1]) << 8) | byteToUnsignedInt(value[0]);
         int humidity = (byteToUnsignedInt(value[3]) << 8) | byteToUnsignedInt(value[2]);
 
-        reading.readings.humidity = (int) ((float) humidity / 100.0f);
-        reading.readings.temperature = (float) temperature / 100.0f;
-        return new Gson().toJson(reading);
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "humidity", "", (int) ((float) humidity / 100.0f)));
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "temperature", "", (float) temperature / 100.0f));
+        return new Gson().toJson(dataPackage);
     }
 
     private static String getMICSensorData(byte[] value) {
-        BleReading reading = new BleReading();
-        reading.readings.noiseLevel = (byteToUnsignedInt(value[1]) << 8) | byteToUnsignedInt(value[0]);
-        return new Gson().toJson(reading);
+        DataPackage dataPackage = new DataPackage();
+        dataPackage.modelId = DeviceModel.MICROPHONE.getId();
+        dataPackage.received = System.currentTimeMillis();
+
+        int noiseLevel = (byteToUnsignedInt(value[1]) << 8) | byteToUnsignedInt(value[0]);
+
+        dataPackage.readings.add(new DataPackage.Data(dataPackage.received, "noiseLevel", "", noiseLevel));
+        return new Gson().toJson(dataPackage);
     }
 
     private static int byteToUnsignedInt(byte b) {