+++ android-sdk/src/main/java/io/relayr/model/TransmitterDevice.java
@@ -32,12 +32,17 @@ public class TransmitterDevice extends Transmitter {
                 o instanceof Device && ((Device) o).id.equals(id);
     }
 
-    public Observable<Object> subscribeToCloudReadings() {
+    public Observable<Reading> subscribeToCloudReadings() {
         return RelayrSdk.getWebSocketClient().subscribe(this);
     }
 
     public void unSubscribeToCloudReadings() {
         RelayrSdk.getWebSocketClient().unSubscribe(id);
+    }
+
+    /** Sends a command to the this device */
+    public Observable<Void> sendCommand(Command command) {
+        return RelayrSdk.getRelayrApi().sendCommand(id, command);
     }
     
 }